requirements.txt
fastapi==0.99.1
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.5
keras==2.8.0
Keras-Preprocessing==1.1.2
numpy==1.22.3
pandas==2.0.3
python-dotenv==1.0.0
torch==2.0.0
requests==2.27.1
streamlit==1.24.0
transformers==4.30.2
uvicorn==0.22.0
